Move has status key into better location
